fix: cast logits to float32 before loss computation to avoid bf16 quantization
bf16 has only 7 mantissa bits, causing observable loss to be quantized
to ~0.03-0.06 resolution. This made 63% of pool samples have zero loss
variance across SGLD draws, and the rest only oscillate between 2
adjacent bf16 levels. The sampler appeared 'leveled off' but was
actually stuck at a single point.

Also includes: discover_checkpoint_dirs fallback for non-standard names,
param_dist computation fix for non-rank-0 chains.

+++ src/bif/training/loss.py
@@ -28,7 +28,7 @@ def per_token_causal_lm_loss(
     Returns:
         Per-token loss tensor of shape (batch, seq_len - 1).
     """
-    log_probs = torch.nn.functional.log_softmax(logits, dim=-1)
+    log_probs = torch.nn.functional.log_softmax(logits.float(), dim=-1)
     shift_log_probs = log_probs[:, :-1, :]
     shift_input_ids = input_ids[:, 1:, None]
     return -shift_log_probs.gather(dim=-1, index=shift_input_ids)[:, :, 0]
@@ -49,7 +49,7 @@ def per_example_causal_lm_loss(
     Returns:
         Per-example loss tensor of shape (batch,).
     """
-    shift_logits = logits[:, :-1, :].contiguous()
+    shift_logits = logits[:, :-1, :].contiguous().float()
     shift_labels = labels[:, 1:].contiguous()
     valid_mask = shift_labels.ne(ignore_index)
     loss_fct = nn.CrossEntropyLoss(reduction="none", ignore_index=ignore_index)
